refactor: Apero Analytic event
