+++ Scripts/ThirdPartyServices/AnalyticEvents/Apero/Purchase.cs
@@ -1,6 +1,6 @@
 ﻿namespace TheOneStudio.UITemplate.UITemplate.ThirdPartyServices.AnalyticEvents.Apero
 {
-    using Core.AnalyticServices.CommonEvents;
+
     using Core.AnalyticServices.Data;
     using ServiceImplementation.IAPServices;
